pyproject.toml
[project]
name = "ble-server"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jijinbei", email = "hallucigenia2002@gmail.com" }
]
dependencies = [
    "requests>=2.32.3",
    "toml>=0.10.2",
    "polars>=1.5.0",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ble_server"]

[tool.rye.scripts]
server = { cmd = "python ./src/ble_server/main.py" }
